\name{app_ui}
\alias{app_ui}
\title{PopPK FitCheckR User Interface}
\usage{
app_ui()
}
\value{
A Shiny UI definition.
}
\description{
Builds the application user interface: the Instructions tab, the Goodness of
Fit (GOF) Plots tab, and the Correlation Plots tab.
}
\seealso{
\code{\link{app_server}}, \code{\link{run_app}}
}
